Development

+++ src/main/java/com/akropoliprishtine/services/economy/PaymentService.java
@@ -2,7 +2,6 @@ package com.akropoliprishtine.services.economy;
 
 
 import com.akropoliprishtine.entities.ApplicationUser;
-import com.akropoliprishtine.entities.Role;
 import com.akropoliprishtine.entities.economy.Payment;
 import com.akropoliprishtine.repositories.UserRepository;
 import com.akropoliprishtine.repositories.economy.PaymentRepository;
@@ -11,39 +10,20 @@ import com.akropoliprishtine.services.EmailService;
 import com.akropoliprishtine.services.JwtUserDetailsService;
 import com.akropoliprishtine.services.RoleService;
 import com.akropoliprishtine.utils.JwtTokenUtil;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
-import javax.swing.text.html.Option;
 import javax.transaction.Transactional;
-import java.util.Date;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
 public class PaymentService {
 
-    @Autowired
     PaymentRepository paymentRepository;
-
-    @Autowired
     UserRepository userRepository;
-
-    @Autowired
-    RoleService roleService;
-
-    @Autowired
     JwtUserDetailsService jwtUserDetailsService;
-
-    @Autowired
     ApplicationUserService userService;
-
-    @Autowired
     EmailService emailService;
-
-    @Autowired
     JwtTokenUtil jwtTokenUtil;
 
     public PaymentService(PaymentRepository paymentRepository,
@@ -70,54 +50,58 @@ public class PaymentService {
             return this.paymentRepository.findAll();
         }
 
-        return this.paymentRepository.findAllByApplicationUser(user);
+        return null;
     }
 
     public Optional<Payment> getPaymentById(Long id) {
-        return this.paymentRepository.findById(id);
+        ApplicationUser user = jwtUserDetailsService.getUserFromToken();
+        String roleName = user.getRole().getName().toLowerCase();
+
+        if ((roleName.contains("pg") && roleName.contains("ekonomia"))
+                || roleName.contains("pishtari")
+                || roleName.contains("kf")) {
+            return this.paymentRepository.findById(id);
+        }
+        return Optional.empty();
     }
 
     @Transactional
     public Payment createPayment(Payment payment) {
-        Payment savedPayment = this.paymentRepository.save(payment);
-        Optional<ApplicationUser> user = this.userService.findById(savedPayment.getApplicationUser().getId());
-        savedPayment.setApplicationUser(user.orElse(null));
-        emailService.sendPaymentEmailToClient(savedPayment);
-        return savedPayment;
+        ApplicationUser appUser = jwtUserDetailsService.getUserFromToken();
+        String roleName = appUser.getRole().getName().toLowerCase();
+
+        if ((roleName.contains("pg") && roleName.contains("ekonomia"))
+                || roleName.contains("pishtari")
+                || roleName.contains("kf")) {
+            Payment savedPayment = this.paymentRepository.save(payment);
+            Optional<ApplicationUser> user = this.userService.findById(savedPayment.getApplicationUser().getId());
+            savedPayment.setApplicationUser(user.orElse(null));
+            emailService.sendPaymentEmailToClient(savedPayment);
+            return savedPayment;
+        }
+        return null;
     }
 
     public Payment updatePayment(Payment payment) {
-        return this.paymentRepository.save(payment);
-    }
+        ApplicationUser appUser = jwtUserDetailsService.getUserFromToken();
+        String roleName = appUser.getRole().getName().toLowerCase();
 
-    public Payment verifyPayment(Long paymentId) {
-        Payment payment = paymentRepository.getOne(paymentId);
-        ApplicationUser user = jwtUserDetailsService.getUserFromToken();
-
-
-        if (user.getId() == payment.getApplicationUser().getId()) {
-            payment.setVerifiedFromUser(true);
-            payment = this.paymentRepository.save(payment);
-
-            Role pg = roleService.findByName("PG tek Ekonomia");
-            Role helper = roleService.findByName("ND tek Ekonomia");
-
-            List<ApplicationUser> pgUsers = userRepository.findAllByRole(pg);
-            List<ApplicationUser> helperUsers = userRepository.findAllByRole(helper);
-
-            pgUsers.forEach(pgUser -> {
-//                emailService.sendBorrowEmailToLibrary(user.getEmail(), payment);
-            });
-
-            helperUsers.forEach(helpUser -> {
-//                emailService.sendBorrowEmailToLibrary(user.getEmail(), payment);
-            });
+        if ((roleName.contains("pg") && roleName.contains("ekonomia"))
+                || roleName.contains("pishtari")
+                || roleName.contains("kf")) {
+            return this.paymentRepository.save(payment);
         }
-
-        return payment;
+        return null;
     }
 
     public void deletePayment(Long id) {
-        this.paymentRepository.deleteById(id);
+        ApplicationUser appUser = jwtUserDetailsService.getUserFromToken();
+        String roleName = appUser.getRole().getName().toLowerCase();
+
+        if ((roleName.contains("pg") && roleName.contains("ekonomia"))
+                || roleName.contains("pishtari")
+                || roleName.contains("kf")) {
+            this.paymentRepository.deleteById(id);
+        }
     }
 }
